# refactor: improve void element parsing perf, fixes #886 and #857