--- a/generator/test-html/template.templ
+++ b/generator/test-html/template.templ
@@ -10,5 +10,5 @@ templ render(p person) {
 	<hr noshade?={ true }/>
 	<hr optionA optionB?={ true } optionC="other" optionD?={ false }/>
 	<hr noshade/>
-	<input name="test">Text</input>
+	<input name="test"/>Text
 }